Order functionality draft
Not tested yet.
Now the item object is still flattened as an array, initially
to hurry the interfacing with the Perl module.
But we were stuck with having to typemap char *** anyway,
so we switch to SWIG because of the potential of ease.
Still, we need to typemap char ***, but now that we're at SWIG,
we might as well typemap a custom item struct.