use_insurance turns out to be required
This one follows the example given for the boolean JSON value,
which is written as true or false.
The response is error 500, internal server error.
Despite the example, it says max length is 1 in the middle part
of the doc, so maybe it should be written as 1 or 0.
Will be tried next.